+++ packages/headless/src/api/search/search-api-params.ts
@@ -150,6 +150,7 @@ export interface PipelineRuleParameters {
 
 export interface GenQAParameters {
   responseFormat: ResponseFormatParameters;
+  citationsFieldToInclude: string[];
 }
 
 export interface ResponseFormatParameters {