+++ packages/quantic/force-app/main/default/lwc/quanticGeneratedAnswer/quanticGeneratedAnswer.js
@@ -23,6 +23,8 @@ import {
 import {AriaLiveRegion, I18nUtils} from 'c/quanticUtils';
 import {LightningElement, api} from 'lwc';
 // @ts-ignore
+import errorTemplate from './templates/errorTemplate.html';
+// @ts-ignore
 import generatedAnswerTemplate from './templates/generatedAnswer.html';
 // @ts-ignore
 import loadingTemplate from './templates/loading.html';
@@ -69,6 +71,13 @@ export default class QuanticGeneratedAnswer extends LightningElement {
    */
   @api answerStyle = 'default';
   /**
+   * A list of fields to fetch with the citations used to generate the answer.
+   * @api
+   * @type {string}
+   * @defaultValue `'sfid,sfkbid,sfkavid'`
+   */
+  @api fieldsToIncludeInCitations = 'sfid,sfkbid,sfkavid';
+  /**
    * Indicates whether footer sections should be displayed on multiple lines.
    * @api
    * @type {boolean}
@@ -109,6 +118,8 @@ export default class QuanticGeneratedAnswer extends LightningElement {
   feedbackSubmitted = false;
   /** @type {import('c/quanticUtils').AriaLiveUtils} */
   ariaLiveMessage;
+  /** @type {boolean} */
+  hasInitializationError = false;
 
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -156,6 +167,7 @@ export default class QuanticGeneratedAnswer extends LightningElement {
         isVisible: storedGeneratedAnswerVisibility === false ? false : true,
         responseFormat: {answerStyle: this.answerStyle},
       },
+      fieldsToIncludeInCitations: this.citationFields,
     });
   }
 
@@ -396,7 +408,24 @@ export default class QuanticGeneratedAnswer extends LightningElement {
     return this.multilineFooter ? false : true;
   }
 
+  get citationFields() {
+    return this.fieldsToIncludeInCitations
+      ?.split(',')
+      .map((field) => field.trim())
+      .filter((field) => field.length > 0);
+  }
+
+  /**
+   * Sets the component in the initialization error state.
+   */
+  setInitializationError() {
+    this.hasInitializationError = true;
+  }
+
   render() {
+    if (this.hasInitializationError) {
+      return errorTemplate;
+    }
     if (this.isLoading) {
       return loadingTemplate;
     }